WIP REDIS refactor 20250901 1833
