+++ crates/zoekt-distributed/src/bin/dzr-admin.rs
@@ -10,8 +10,6 @@ use axum::{
 };
 use chrono::{DateTime, TimeZone, Utc};
 use clap::Parser;
-use deadpool_redis::Config as RedisConfig;
-// hmac handled in web_utils now when building cookies
 use parking_lot::RwLock;
 use serde::Deserialize;
 use serde_json::json;
@@ -1597,61 +1595,8 @@ async fn main() -> Result<()> {
     )?;
 
     // Build redis pool from REDIS_URL if present, with optional authentication
-    let redis_pool = match std::env::var("REDIS_URL").ok() {
-        Some(mut url) => {
-            // Check if URL already has authentication
-            let has_auth = url.contains('@');
-
-            // If no auth in URL, try to add it from environment variables
-            if !has_auth {
-                if let Ok(username) = std::env::var("REDIS_USERNAME") {
-                    if let Ok(password) = std::env::var("REDIS_PASSWORD") {
-                        // Insert username:password@ after redis://
-                        if let Some(pos) = url.find("://") {
-                            let auth_part = format!("{}:{}@", username, password);
-                            url.insert_str(pos + 3, &auth_part);
-                        }
-                    } else {
-                        // Username without password
-                        if let Some(pos) = url.find("://") {
-                            let auth_part = format!("{}@", username);
-                            url.insert_str(pos + 3, &auth_part);
-                        }
-                    }
-                } else if let Ok(password) = std::env::var("REDIS_PASSWORD") {
-                    // Password without username
-                    if let Some(pos) = url.find("://") {
-                        let auth_part = format!(":{}@", password);
-                        url.insert_str(pos + 3, &auth_part);
-                    }
-                }
-            }
-
-            RedisConfig::from_url(&url)
-                .create_pool(None)
-                .ok()
-                .map(|p| std::sync::Arc::new(RealRedis { pool: p }) as std::sync::Arc<dyn DynRedis>)
-        }
-        None => {
-            // No REDIS_URL provided, but check if we have auth credentials to construct one
-            if let (Ok(username), Ok(password)) = (
-                std::env::var("REDIS_USERNAME"),
-                std::env::var("REDIS_PASSWORD"),
-            ) {
-                let url = format!("redis://{}:{}@127.0.0.1:6379", username, password);
-                RedisConfig::from_url(&url).create_pool(None).ok().map(|p| {
-                    std::sync::Arc::new(RealRedis { pool: p }) as std::sync::Arc<dyn DynRedis>
-                })
-            } else if let Ok(password) = std::env::var("REDIS_PASSWORD") {
-                let url = format!("redis://:{}@127.0.0.1:6379", password);
-                RedisConfig::from_url(&url).create_pool(None).ok().map(|p| {
-                    std::sync::Arc::new(RealRedis { pool: p }) as std::sync::Arc<dyn DynRedis>
-                })
-            } else {
-                None
-            }
-        }
-    };
+    let redis_pool = zoekt_distributed::redis_adapter::create_redis_pool()
+        .map(|p| std::sync::Arc::new(RealRedis { pool: p }) as std::sync::Arc<dyn DynRedis>);
 
     // Admin credentials from env (user requested ZOEKT_ADMIN_{USERNAME,PASSWORD})
     let admin_user = std::env::var("ZOEKT_ADMIN_USERNAME").unwrap_or_else(|_| "admin".into());
